# components/tenant-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.CreateFederationTokenRequest:
    properties:
      description:
        maxLength: 256
        type: string
      tenant_id:
        maxLength: 128
        minLength: 1
        type: string
    required:
    - tenant_id
    type: object
  internal_handler.CreateFederationTokenResponse:
    properties:
      record:
        $ref: '#/definitions/internal_handler.FederationTokenRecord'
      token:
        type: string
    type: object
...
  internal_handler.FederationTokenRecord:
    properties:
      description:
        type: string
      expires_at:
        type: string
      issued_at:
        type: string
      issued_by:
        type: string
      tenant_id:
        type: string
      token_id:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Multi-tenant configuration management API for the Dynamic Alerting Platform.
    Provides CRUD over per-tenant alert thresholds, RBAC-scoped writes, and async batch operations.
    Auth is enforced by the oauth2-proxy sidecar and forwarded via X-Forwarded-Email / X-Forwarded-Groups.
  license:
    name: See repository LICENSE
  title: Tenant API
  version: v2.8.0
paths:
...
  /api/v1/federation/tokens:
    get:
      description: Returns the non-expired federation token records for the tenant
        named by the tenant_id query parameter. Requires admin permission on that
        tenant. The signed JWTs themselves are not returned.
      parameters:
      - description: Tenant ID
        in: query
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handler.FederationTokenRecord'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List a tenant's federation tokens
      tags:
      - federation
    post:
      consumes:
      - application/json
      description: Mints a short-lived RS256 JWT for the named tenant (ADR-020). Requires
        admin permission on the tenant. The signed token is returned once and is not
        retrievable afterwards.
      parameters:
      - description: Token request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateFederationTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.CreateFederationTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Issue a tenant federation token
      tags:
      - federation
  /api/v1/federation/tokens/{id}:
    delete:
      description: 'Revokes a federation token — removes its record and adds it to
        the gateway revoked set. NOTE: revocation is eventually consistent and propagates
        to the gateway within ~1-2 minutes (ADR-020 Posture B). Requires admin permission
        on the token''s tenant.'
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Revoke a federation token
      tags:
      - federation
...
swagger: "2.0"